added missing classes from the external client response
added timeouts for the external client calls

implemented tests for the validator and converter

minor improvements on the json response

+++ src/main/java/com/lasiqueira/weatherforecast/api/dto/v1/WeatherForecastMetricsDTO.java
@@ -3,10 +3,12 @@ package com.lasiqueira.weatherforecast.api.dto.v1;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Builder
 @Data
 public class WeatherForecastMetricsDTO {
-    private final Double averageTemperatureDay;
-    private final Double averageTemperatureNight;
-    private final Double averagePressure;
+    private final BigDecimal averageTemperatureDay;
+    private final BigDecimal averageTemperatureNight;
+    private final BigDecimal averagePressure;
 }
