+++ src/main/java/com/lasiqueira/weatherforecast/model/WeatherForecastMetrics.java
@@ -3,10 +3,12 @@ package com.lasiqueira.weatherforecast.model;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Builder
 @Data
 public class WeatherForecastMetrics {
-    private final Double averageTemperatureDay;
-    private final Double averageTemperatureNight;
-    private final Double averagePressure;
+    private final BigDecimal averageTemperatureDay;
+    private final BigDecimal averageTemperatureNight;
+    private final BigDecimal averagePressure;
 }
